requirement.txt
chromadb
sentence-transformers
numpy
fitz
pytesseract
Pillow
requests
beautifulsoup4
ollama
